separate cgu and dpa in separate concerns
